text search and more keywords

+++ update-db.py
@@ -8,11 +8,11 @@ import subprocess
 import concurrent.futures
 
 sys.path.append('sidneycadot/oeis')
-from sidneycadot.oeis.oeis_entry    import parse_oeis_entry
-from sidneycadot.oeis.timer         import start_timer
-from sidneycadot.oeis.exit_scope    import close_when_done
-from sidneycadot.oeis.setup_logging import setup_logging
 import sidneycadot.oeis.fetch_oeis_database as fetch_oeis_database
+from sidneycadot.oeis.setup_logging import setup_logging
+from sidneycadot.oeis.exit_scope import close_when_done
+from sidneycadot.oeis.timer import start_timer
+from sidneycadot.oeis.oeis_entry import parse_oeis_entry
 
 
 logger = logging.getLogger(__name__)
@@ -22,7 +22,8 @@ def create_database_schema(dbconn):
              CREATE TABLE IF NOT EXISTS seq_entries (
                  oeis_id               INTEGER  PRIMARY KEY NOT NULL,
                  name                  TEXT     NOT NULL,
-                 keywords              TEXT     NOT NULL
+                 keywords              TEXT     NOT NULL,
+                 contributors          TEXT     NOT NULL
              );
              """
     dbconn.execute(schema)
@@ -30,17 +31,33 @@ def create_database_schema(dbconn):
 
 def process_oeis_entry(oeis_entry):
     (oeis_id, main_content, bfile_content) = oeis_entry
-    parsed_entry = parse_oeis_entry(oeis_id, main_content, bfile_content)
-    keywords = parsed_entry.keywords
-    prefix = int(parsed_entry.oeis_id/1000)
-    if os.path.exists('/tmp/loda-programs/oeis/{:03}/A{:06}.asm'.format(prefix, parsed_entry.oeis_id)):
-        keywords.append('loda')
-    if os.path.exists('/tmp/joeis/src/irvine/oeis/a{:03}/A{:06}.java'.format(prefix, parsed_entry.oeis_id)):
+    entry = parse_oeis_entry(oeis_id, main_content, bfile_content)
+    keywords = entry.keywords
+    contributors = []
+    if entry.author and len(entry.author) > 0:
+        contributors.append(entry.author.replace("_", ""))
+    if entry.maple_programs and len(entry.maple_programs) > 0:
+        keywords.append('maple')
+    if entry.mathematica_programs and len(entry.mathematica_programs) > 0:
+        keywords.append('mathematica')
+    if entry.other_programs and entry.other_programs.find('(PARI)') >= 0:
+        keywords.append('pari')
+    prefix = int(entry.oeis_id/1000)
+    java_path = '/tmp/joeis/src/irvine/oeis/a{:03}/A{:06}.java'.format(prefix, entry.oeis_id)
+    if os.path.exists(java_path):
         keywords.append('java')
+    loda_path = os.path.join(os.path.expanduser('~'), 'loda', 'programs', 'oeis', '{:03}'.format(prefix), 'A{:06}.asm'.format(entry.oeis_id))
+    if os.path.exists(loda_path):
+        keywords.append('loda')
+        with open(loda_path) as loda_file:
+            for line in loda_file:
+                if line.startswith('; Submitted by '):
+                    contributors.append(line[14:].strip())
     result = (
-        parsed_entry.oeis_id,
-        parsed_entry.name,
-        ",".join(str(keyword) for keyword in parsed_entry.keywords)
+        entry.oeis_id,
+        entry.name,
+        ','.join(str(k) for k in keywords),
+        ','.join(str(c) for c in contributors)
     )
     return result
 
@@ -52,16 +69,20 @@ def process_database_entries(file_in, file_out):
             with close_when_done(sqlite3.connect(file_out)) as dbconn_out, close_when_done(dbconn_out.cursor()) as dbcursor_out:
                 create_database_schema(dbconn_out)
                 with concurrent.futures.ProcessPoolExecutor() as pool:
-                    dbcursor_in.execute("SELECT oeis_id, main_content, bfile_content FROM oeis_entries ORDER BY oeis_id;")
+                    dbcursor_in.execute(
+                        "SELECT oeis_id, main_content, bfile_content FROM oeis_entries ORDER BY oeis_id;")
                     while True:
                         oeis_entries = dbcursor_in.fetchmany(BATCH_SIZE)
                         if len(oeis_entries) == 0:
                             break
-                        logger.log(logging.PROGRESS, "Processing OEIS entries A{:06} to A{:06} ...".format(oeis_entries[0][0], oeis_entries[-1][0]))
-                        query = "INSERT INTO seq_entries(oeis_id, name, keywords) VALUES (?, ?, ?);"
-                        dbcursor_out.executemany(query, pool.map(process_oeis_entry, oeis_entries))
+                        logger.log(logging.PROGRESS, "Processing OEIS entries A{:06} to A{:06} ...".format(
+                            oeis_entries[0][0], oeis_entries[-1][0]))
+                        query = "INSERT INTO seq_entries(oeis_id, name, keywords, contributors) VALUES (?, ?, ?, ?);"
+                        dbcursor_out.executemany(query, pool.map(
+                            process_oeis_entry, oeis_entries))
                         dbconn_out.commit()
-        logger.info("Processed all database entries in {}.".format(timer.duration_string()))
+        logger.info("Processed all database entries in {}.".format(
+            timer.duration_string()))
 
 
 def update_repo(url, path):
@@ -76,11 +97,22 @@ def update_repo(url, path):
 
 
 def main():
-    update_repo('https://github.com/loda-lang/loda-programs.git', '/tmp/loda-programs')
+    # update joeis
     update_repo('https://github.com/archmageirvine/joeis.git', '/tmp/joeis')
+
+    # update loda
+    loda_bin = os.path.join(os.path.expanduser('~'), 'loda', 'bin', 'loda')
+    p = subprocess.run([loda_bin, 'update'])
+    if p.returncode != 0:
+        logger.error('error updating loda')
+        exit(1)
+
+    # update oeis
     os.makedirs('logfiles', exist_ok=True)
     fetch_oeis_database.main()
-    with setup_logging('logfiles/oeis_parsed.log'):
+
+    # generate final database
+    with setup_logging('logfiles/seqs_parsed.log'):
         if os.path.exists('seqs.sqlite3'):
             os.remove('seqs.sqlite3')
         process_database_entries('oeis.sqlite3', 'seqs.sqlite3')
